# api/school/v1/teacher_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type TeacherLogicer interface {
	Create(ctx context.Context, req *CreateTeacherRequest) (*CreateTeacherReply, error)
	DeleteByID(ctx context.Context, req *DeleteTeacherByIDRequest) (*DeleteTeacherByIDReply, error)
	UpdateByID(ctx context.Context, req *UpdateTeacherByIDRequest) (*UpdateTeacherByIDReply, error)
	GetByID(ctx context.Context, req *GetTeacherByIDRequest) (*GetTeacherByIDReply, error)
	List(ctx context.Context, req *ListTeacherRequest) (*ListTeacherReply, error)
	DeleteByIDs(ctx context.Context, req *DeleteTeacherByIDsRequest) (*DeleteTeacherByIDsReply, error)
	GetByCondition(ctx context.Context, req *GetTeacherByConditionRequest) (*GetTeacherByConditionReply, error)
	ListByIDs(ctx context.Context, req *ListTeacherByIDsRequest) (*ListTeacherByIDsReply, error)
	ListByLastID(ctx context.Context, req *ListTeacherByLastIDRequest) (*ListTeacherByLastIDReply, error)
}

type TeacherOption func(*teacherOptions)

type teacherOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *teacherOptions) apply(opts ...TeacherOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithTeacherHTTPResponse() TeacherOption {
	return func(o *teacherOptions) {
		o.isFromRPC = false
	}
}

func WithTeacherRPCResponse() TeacherOption {
	return func(o *teacherOptions) {
		o.isFromRPC = true
	}
}

func WithTeacherResponser(responser errcode.Responser) TeacherOption {
	return func(o *teacherOptions) {
		o.responser = responser
	}
}

func WithTeacherLogger(zapLog *zap.Logger) TeacherOption {
	return func(o *teacherOptions) {
		o.zapLog = zapLog
	}
}

func WithTeacherErrorToHTTPCode(e ...*errcode.Error) TeacherOption {
	return func(o *teacherOptions) {
		o.httpErrors = e
	}
}

func WithTeacherRPCStatusToHTTPCode(s ...*errcode.RPCStatus) TeacherOption {
	return func(o *teacherOptions) {
		o.rpcStatus = s
	}
}

func WithTeacherWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) TeacherOption {
	return func(o *teacherOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterTeacherRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic TeacherLogicer,
	opts ...TeacherOption) {

	o := &teacherOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &teacherRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type teacherRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                TeacherLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *teacherRouter) register() {
	r.iRouter.Handle("POST", "/api/v1/teacher", r.withMiddleware("POST", "/api/v1/teacher", r.Create_0)...)
	r.iRouter.Handle("DELETE", "/api/v1/teacher/:id", r.withMiddleware("DELETE", "/api/v1/teacher/:id", r.DeleteByID_0)...)
	r.iRouter.Handle("PUT", "/api/v1/teacher/:id", r.withMiddleware("PUT", "/api/v1/teacher/:id", r.UpdateByID_0)...)
	r.iRouter.Handle("GET", "/api/v1/teacher/:id", r.withMiddleware("GET", "/api/v1/teacher/:id", r.GetByID_0)...)
	r.iRouter.Handle("POST", "/api/v1/teacher/list", r.withMiddleware("POST", "/api/v1/teacher/list", r.List_0)...)
	r.iRouter.Handle("POST", "/api/v1/teacher/delete/ids", r.withMiddleware("POST", "/api/v1/teacher/delete/ids", r.DeleteByIDs_0)...)
	r.iRouter.Handle("POST", "/api/v1/teacher/condition", r.withMiddleware("POST", "/api/v1/teacher/condition", r.GetByCondition_0)...)
	r.iRouter.Handle("POST", "/api/v1/teacher/list/ids", r.withMiddleware("POST", "/api/v1/teacher/list/ids", r.ListByIDs_0)...)
	r.iRouter.Handle("GET", "/api/v1/teacher/list", r.withMiddleware("GET", "/api/v1/teacher/list", r.ListByLastID_0)...)

}

func (r *teacherRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *teacherRouter) Create_0(c *gin.Context) {
	req := &CreateTeacherRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.Create(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) DeleteByID_0(c *gin.Context) {
	req := &DeleteTeacherByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) UpdateByID_0(c *gin.Context) {
	req := &UpdateTeacherByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) GetByID_0(c *gin.Context) {
	req := &GetTeacherByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) List_0(c *gin.Context) {
	req := &ListTeacherRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.List(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) DeleteByIDs_0(c *gin.Context) {
	req := &DeleteTeacherByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByIDs(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) GetByCondition_0(c *gin.Context) {
	req := &GetTeacherByConditionRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByCondition(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) ListByIDs_0(c *gin.Context) {
	req := &ListTeacherByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByIDs(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *teacherRouter) ListByLastID_0(c *gin.Context) {
	req := &ListTeacherByLastIDRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByLastID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}