# docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "user api docs",
    "version": "2.0"
  },
  "tags": [
    {
      "name": "user"
    },
    {
      "name": "teacher"
    }
  ],
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/hello": {
      "get": {
        "summary": "hello",
        "description": "hello",
        "operationId": "user_Hello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1HelloReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/api/v1/login": {
      "post": {
        "summary": "登录",
        "description": "使用邮箱登录",
        "operationId": "user_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LoginReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginRequest"
            }
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/api/v1/teacher": {
      "post": {
        "summary": "create teacher",
        "operationId": "teacher_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateTeacherReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateTeacherRequest"
            }
          }
        ],
        "tags": [
          "teacher"
        ]
      }
    },
    "/api/v1/teacher/condition": {
      "post": {
        "summary": "get teacher by condition",
        "operationId": "teacher_GetByCondition",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetTeacherByConditionReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetTeacherByConditionRequest"
            }
          }
        ],
        "tags": [
          "teacher"
        ]
      }
    },
    "/api/v1/teacher/delete/ids": {
      "post": {
        "summary": "delete teacher by batch id",
        "operationId": "teacher_DeleteByIDs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteTeacherByIDsReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1DeleteTeacherByIDsRequest"
            }
          }
        ],
        "tags": [
          "teacher"
        ]
      }
    },
    "/api/v1/teacher/list": {
      "get": {
        "summary": "list teacher by last id",
        "operationId": "teacher_ListByLastID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListTeacherByLastIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "lastID",
            "description": "last id",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "description": "limit size per page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "sort",
            "description": "sort by column name of table, default is -id, the - sign indicates descending order.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "teacher"
        ]
      },
      "post": {
        "summary": "list of teacher by query parameters",
        "operationId": "teacher_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListTeacherReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListTeacherRequest"
            }
          }
        ],
        "tags": [
          "teacher"
        ]
      }
    },
    "/api/v1/teacher/list/ids": {
      "post": {
        "summary": "list of teacher by batch id",
        "operationId": "teacher_ListByIDs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListTeacherByIDsReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListTeacherByIDsRequest"
            }
          }
        ],
        "tags": [
          "teacher"
        ]
      }
    },
    "/api/v1/teacher/{id}": {
      "get": {
        "summary": "get teacher by id",
        "operationId": "teacher_GetByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetTeacherByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "teacher"
        ]
      },
      "delete": {
        "summary": "delete teacher by id",
        "operationId": "teacher_DeleteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteTeacherByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "teacher"
        ]
      },
      "put": {
        "summary": "update teacher by id",
        "operationId": "teacher_UpdateByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateTeacherByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/teacherUpdateByIDBody"
            }
          }
        ],
        "tags": [
          "teacher"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "teacherUpdateByIDBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "bornAt": {
          "type": "string"
        },
        "job": {
          "$ref": "#/definitions/v1Job"
        }
      }
    },
    "typesColumn": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "column name"
        },
        "exp": {
          "type": "string",
          "title": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like, in"
        },
        "value": {
          "type": "string",
          "title": "column value"
        },
        "logic": {
          "type": "string",
          "title": "logical type, defaults to and when value is null, only \u0026(and), ||(or)"
        }
      }
    },
    "typesConditions": {
      "type": "object",
      "properties": {
        "columns": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/typesColumn"
          },
          "title": "query conditions"
        }
      }
    },
    "typesParams": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32",
          "title": "page number, starting from 0"
        },
        "limit": {
          "type": "integer",
          "format": "int32",
          "title": "number per page"
        },
        "sort": {
          "type": "string",
          "title": "sorted fields, multi-column sorting separated by commas"
        },
        "columns": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/typesColumn"
          },
          "title": "query conditions"
        }
      }
    },
    "v1CreateTeacherReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "v1CreateTeacherRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "bornAt": {
          "type": "string"
        },
        "job": {
          "$ref": "#/definitions/v1Job"
        }
      }
    },
    "v1DeleteTeacherByIDReply": {
      "type": "object"
    },
    "v1DeleteTeacherByIDsReply": {
      "type": "object"
    },
    "v1DeleteTeacherByIDsRequest": {
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1GetTeacherByConditionReply": {
      "type": "object",
      "properties": {
        "teacher": {
          "$ref": "#/definitions/v1Teacher"
        }
      }
    },
    "v1GetTeacherByConditionRequest": {
      "type": "object",
      "properties": {
        "conditions": {
          "$ref": "#/definitions/typesConditions"
        }
      }
    },
    "v1GetTeacherByIDReply": {
      "type": "object",
      "properties": {
        "teacher": {
          "$ref": "#/definitions/v1Teacher"
        }
      }
    },
    "v1HelloReply": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string"
        }
      }
    },
    "v1Job": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "salary": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1ListTeacherByIDsReply": {
      "type": "object",
      "properties": {
        "teachers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Teacher"
          }
        }
      }
    },
    "v1ListTeacherByIDsRequest": {
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1ListTeacherByLastIDReply": {
      "type": "object",
      "properties": {
        "teachers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Teacher"
          }
        }
      }
    },
    "v1ListTeacherReply": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int64"
        },
        "teachers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Teacher"
          }
        }
      }
    },
    "v1ListTeacherRequest": {
      "type": "object",
      "properties": {
        "params": {
          "$ref": "#/definitions/typesParams"
        }
      }
    },
    "v1LoginReply": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "v1Teacher": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "bornAt": {
          "type": "string"
        },
        "job": {
          "$ref": "#/definitions/v1Job"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "v1UpdateTeacherByIDReply": {
      "type": "object"
    }
  }
}